feat(ConvertHandler, api main, Unit Tests suite): :fire: transfer inputNum parsing to ConvertHandler, added 6th fcc test & relevant code
added class syntactic sugar to ConvertHandler, and defined methods as static. Implemented test to default to a numeric value of 1 if no client numerical imput is provided, and code to implement feature in the relevant ConvertHandler method getNum. Moved local conversion variables to suite scope, and renamed according to static variable convention.