feat: :sparkles: added test and code for POST requests with every field included
added the following utility functions to the assert function in issue-tracker.test.js; areNull, areStrings, areObjects, and areTrue. Routed post requests in the router to the handler postProjectRequest method. Corrected JSDocs in some handler methods and added the postProjectRequest method. Refactored collection from injectDB to a global variable in issues-dao, and edited error messages and error objects that are returned. Added the postProject method with documentation, and added @typedefs to issues-dao.