feat: :adhesive_bandage: added code to pass test #3
IssueHandler post method (previously postProjectRequest) now responds with a bad request status if any of the required properties are missing in a project's issue, and sets any missing optional properties to null instead of leaving them as undefined, for consistency purposes. Also renamed putProjectRequest and getProjectRequest as put and get respectively, and dropTestRequest as drop.