test: :white_check_mark: GET request with multiple url params (filters)
added GET test with multiple url params. In the IssuesDAO, instead of searching for exact matches with ===, we search for null IF the query is null; if the query is a string, we assert the respective issue field MATCHES a RegExp instance with the query string passed (case-insensitive). Use a RegExp instance instead of the string includes method precisely to add case-insensitivity (with the 'i' flag).