feat: :sparkles: added api checking for _id, and moved data handling to the Handler
moved data handling (issue filtering, returning issues instead of the whole project doc, etc.) to the Handler so that the DAO only makes db requests and handles server errors. Made post requests return error messages if the document sent contains an _id, as this is supplied by the handler. Thought overwriting this property might lead to uenxpected behaviour, so it's best to not go through with it unless the posted document contains precisely the properties required (and/or the optional properties, too).