test: :test_tube: added POST comment test to test #1, instead of creating test #4
since a GET request is required to fetch the _id of the book to append the comment to, implemented POST comment test in addition to test #1, instead of making a new test that makes a GET request identical to test #1's. Updated tasks in dev notes. Fixed bug in personal-library.js (arrow function with "this" keyword). Created Handler addComment method, which is linked to the /api/books/:_id endpoint.